Change recording and tickcount type to be uint32_t so it's the same on linux.

+++ speed-dreams/source-2.2.3/src/simulatedDrivingAssistance/backend/Mediator.h
@@ -109,7 +109,7 @@ private:
 
     DecisionMaker m_decisionMaker;
 
-    unsigned long m_tickCount = 0;
+    uint32_t m_tickCount = 0;
     tDatabaseSettings m_dbSettings;
 
     tTrack* m_track = nullptr;
